Insert Table
Built the initial table records - double check my numbers to be sure my 'syntax' is correct,
--all person are input
--25 records in our 4 parts input
--10 loans input that match 10 of the records

+++ includes/data/create_table.sql
@@ -82,7 +82,7 @@ CREATE TABLE BIB_TECH (
     BibTech_Stock varchar(25),
     BibTech_Edgecode_Date year,
     BibTech_Footage int,
-    BibTech_Can_Size smallint,
+    BibTech_Can_Size int,
     primary key (BibTech_ID), 
     foreign key (Bib_IU_Barcode) references BIB_BASIC on update cascade
 );
